+++ src/command/sample_encode.rs
@@ -41,7 +41,7 @@ pub struct Args {
 
     /// Encoder constant rate factor (1-63). Lower means better quality.
     #[arg(long)]
-    pub crf: u8,
+    pub crf: f32,
 
     #[clap(flatten)]
     pub sample: args::Sample,
